SQL stuff is now in server_util.c
